+++ CHANGELOG.md
@@ -1,7 +1,8 @@
-# 0.2.0 2020/11/19
+# 0.2.1 2020/11/19
 
-- reduce download size by delegate qrcode generation via saihu qrcode
-- fix i18n issue
+- reduce download size by
+  - move i18n and addon.exec to saihu 0.34
+  - delegate qrcode generation via saihu qrcode
 
 # 0.1.0 2020/11/18